Move utility types from resolver into resolver::types
